+++ rational/tests/simplify.rs
@@ -4,22 +4,22 @@ use dashu_ratio::RBig;
 mod helper_macros;
 
 #[test]
-fn test_simpliest_in() {
-    assert_eq!(RBig::simpliest_in(rbig!(0), rbig!(0)), rbig!(0));
-    assert_eq!(RBig::simpliest_in(rbig!(-1), rbig!(1)), rbig!(0));
-    assert_eq!(RBig::simpliest_in(rbig!(-1), rbig!(-1)), rbig!(-1));
-    assert_eq!(RBig::simpliest_in(rbig!(2/7), rbig!(2/9)), rbig!(1/4));
-    assert_eq!(RBig::simpliest_in(rbig!(-20/7), rbig!(-20/9)), rbig!(-5/2));
-    assert_eq!(RBig::simpliest_in(rbig!(5), rbig!(7)), rbig!(6));
-    assert_eq!(RBig::simpliest_in(rbig!(5), rbig!(6)), rbig!(11/2));
+fn test_simplest_in() {
+    assert_eq!(RBig::simplest_in(rbig!(0), rbig!(0)), rbig!(0));
+    assert_eq!(RBig::simplest_in(rbig!(-1), rbig!(1)), rbig!(0));
+    assert_eq!(RBig::simplest_in(rbig!(-1), rbig!(-1)), rbig!(-1));
+    assert_eq!(RBig::simplest_in(rbig!(2/7), rbig!(2/9)), rbig!(1/4));
+    assert_eq!(RBig::simplest_in(rbig!(-20/7), rbig!(-20/9)), rbig!(-5/2));
+    assert_eq!(RBig::simplest_in(rbig!(5), rbig!(7)), rbig!(6));
+    assert_eq!(RBig::simplest_in(rbig!(5), rbig!(6)), rbig!(11/2));
 }
 
 #[test]
-fn test_simpliest_from_f32() {
-    assert_eq!(RBig::simpliest_from_f32(0f32), Some(rbig!(0)));
-    assert_eq!(RBig::simpliest_from_f32(f32::INFINITY), None);
-    assert_eq!(RBig::simpliest_from_f32(f32::NEG_INFINITY), None);
-    assert_eq!(RBig::simpliest_from_f32(f32::NAN), None);
+fn test_simplest_from_f32() {
+    assert_eq!(RBig::simplest_from_f32(0f32), Some(rbig!(0)));
+    assert_eq!(RBig::simplest_from_f32(f32::INFINITY), None);
+    assert_eq!(RBig::simplest_from_f32(f32::NEG_INFINITY), None);
+    assert_eq!(RBig::simplest_from_f32(f32::NAN), None);
 
     let cases = [
         // (numerator, denominator)
@@ -39,16 +39,16 @@ fn test_simpliest_from_f32() {
     ];
     for (num, den) in cases {
         let f = num as f32 / den as f32;
-        assert_eq!(RBig::simpliest_from_f32(f).unwrap(), RBig::from_parts(num.into(), den.into()));
+        assert_eq!(RBig::simplest_from_f32(f).unwrap(), RBig::from_parts(num.into(), den.into()));
     }
 }
 
 #[test]
-fn test_simpliest_from_f64() {
-    assert_eq!(RBig::simpliest_from_f64(0f64), Some(rbig!(0)));
-    assert_eq!(RBig::simpliest_from_f64(f64::INFINITY), None);
-    assert_eq!(RBig::simpliest_from_f64(f64::NEG_INFINITY), None);
-    assert_eq!(RBig::simpliest_from_f64(f64::NAN), None);
+fn test_simplest_from_f64() {
+    assert_eq!(RBig::simplest_from_f64(0f64), Some(rbig!(0)));
+    assert_eq!(RBig::simplest_from_f64(f64::INFINITY), None);
+    assert_eq!(RBig::simplest_from_f64(f64::NEG_INFINITY), None);
+    assert_eq!(RBig::simplest_from_f64(f64::NAN), None);
 
     let cases = [
         // (numerator, denominator)
@@ -79,7 +79,7 @@ fn test_simpliest_from_f64() {
     ];
     for (num, den) in cases {
         let f = num as f64 / den as f64;
-        assert_eq!(RBig::simpliest_from_f64(f).unwrap(), RBig::from_parts(num.into(), den.into()));
+        assert_eq!(RBig::simplest_from_f64(f).unwrap(), RBig::from_parts(num.into(), den.into()));
     }
 }