+++ integer/CHANGELOG.md
@@ -6,12 +6,14 @@
 
 - Implement `Gcd::gcd` and `ExtendedGcd::gcd_ext` between `UBig` and `IBig`
 - Implement `DivRem::div_rem` between `UBig` and `IBig`
-- Add `abs_bit_len` for `IBig`
+- Add `bit_len` for `IBig`
 
 ### Change
 
 - `Modulo::inv` is renamed to `Modulo::inverse`
-- `sqrt_rem` is only exposed through the `RootRem` trait now.
+- `sqrt_rem` is only exposed through the `dashu_base::RootRem` trait now.
+- `abs_cmp` is only exposed throught the `dashu_base::AbsCmp` trait now.
+- `abs_eq` is only exposed throught the `dashu_base::AbsEq` trait now.
 
 ### Remove